bugfix, now seems to work

+++ cluster.py
@@ -1,21 +1,10 @@
 __author__ = 'hofmann'
 
 
-#!/bin/bash
-#HOME_DIR="/home/user03/pyprojects/taxonomic_classification_plus_otu/"
-#t1="/home/user03/output/nobackup/2014_09_24_random1000_ref_silva_hmmer3/16S_rRNA.fna"
-#t2="/home/user03/output/nobackup/2014_09_24_random1000_ref_silva_hmmer3/mothur_cluster_list.txt"
-#t3="/home/user03/pyprojects/taxonomic_classification_plus_otu/references/"
-#t4="0.05"
-#t5="no longer used"
-#t6="45"
-
-
 import os
 import glob
 import shutil
 import tempfile
-
 from source.argumenthandler import ArgumentHandler
 from source.logger import Logger
 
@@ -26,9 +15,11 @@ def main(options):
 						   directory_silva_reference=options.silva_reference_directory,
 						   max_processors=options.processors,
 						   debug=options._debug_mode)
+
 	return mg_cluster.cluster(marker_gene_fasta=os.path.join(options.project_directory, options.file_mg_16s),
-						output_cluster_file=options.processors,
-						distance_cutoff=options.distance_cutoff)
+						output_cluster_file=os.path.join(options.project_directory, options.file_cluster_mg_16s),
+						distance_cutoff=options.distance_cutoff,
+						precision=1000)
 
 
 class MGCluster(object):
@@ -38,61 +29,84 @@ class MGCluster(object):
 			self._logger = Logger("MGCluster")
 		self._mothur_executable = mothur_executable
 		self._working_dir = tempfile.mkdtemp()
+
+		self._old_dir = os.getcwd()
+		# required or mothur messes up the dist.seqs command
+		# do not use absolut paths!!!
+		os.chdir(self._working_dir)
+
 		self._max_processors = max_processors
 		self._debug = debug
-		self._ref_silva_distances = os.path.join(directory_silva_reference, "mothur_ref_distances")
-		self._ref_silva_names = os.path.join(directory_silva_reference, "mothur_ref_names")  # unique
-		self._ref_silva_alignment = os.path.join(directory_silva_reference, "mothur_alignment_ref.fasta")
-		self._local_distance = os.path.join(self._working_dir, "ref.align.dist")
+		ref_silva_distances = self._get_symbolic_link_path(os.path.join(directory_silva_reference, "mothur_ref_distances"))
+		ref_silva_names = self._get_symbolic_link_path(os.path.join(directory_silva_reference, "mothur_ref_names"))  # unique
+		ref_silva_alignment = self._get_symbolic_link_path(os.path.join(directory_silva_reference, "mothur_alignment_ref.fasta"))
+		self._ref_silva_distances = ref_silva_distances
+		self._ref_silva_names = ref_silva_names
+		self._ref_silva_alignment = ref_silva_alignment
+		local_distance = os.path.join(self._working_dir, "ref.align.dist")
+		self._local_distance = "ref.align.dist"
 
-	def cluster(self, marker_gene_fasta, output_cluster_file, distance_cutoff):
+	def cluster(self, marker_gene_fasta, output_cluster_file, distance_cutoff, precision):
 		local_marker_gene_fasta = self._get_symbolic_link_path(marker_gene_fasta)
 		shutil.copy2(self._ref_silva_distances, self._local_distance)
+		shutil.copy2(self._ref_silva_distances, self._local_distance+"_")
+		#os.system("touch {file}".format(file=self._local_distance))
 
-		mothur_cmd = self._get_mothur_cmd(local_marker_gene_fasta, distance_cutoff)
-		cmd = "echo -e \"{mothur_cmd}\" | {mothur_executable}".format(mothur_cmd=mothur_cmd,
-																	  mothur_executable=self._mothur_executable)
+		mothur_cmd = self._get_mothur_cmd(local_marker_gene_fasta, distance_cutoff, precision)
+		cmd = "echo \"{mothur_cmd}\" | {mothur_executable}".format(mothur_cmd=mothur_cmd,
+																   mothur_executable=self._mothur_executable)
 		os.system(cmd)
+		os.chdir(self._old_dir)
 
 		find_mask_list = os.path.join(self._working_dir, "*.list")
 		list_of_files = glob.glob(find_mask_list)
 		if len(list_of_files) == 0:
 			self._logger.error("[MGCluster] Clustering failed")
-			self._logger.warning("[MGCluster] Remove manually: '{}".format(self._working_dir))
+			self._logger.warning("[MGCluster] Remove manually: '{}'".format(self._working_dir))
 			return False
 		elif len(list_of_files) == 1:
-			shutil.copy2(list_of_files[0], output_cluster_file)
+			local_distance = os.path.join(self._working_dir, "ref.align.dist")
+			if os.path.exists(local_distance):
+				shutil.copy2(list_of_files[0], output_cluster_file)
+				self._logger.info("[MGCluster] Clustering success")
+			else:
+				self._logger.error("[MGCluster] Clustering failed!")
 			if not self._debug:
 				shutil.rmtree(self._working_dir)
-				self._logger.warning("[MGCluster] Remove manually: '{}".format(self._working_dir))
-			self._logger.error("[MGCluster] Clustering success")
+			else:
+				self._logger.warning("[MGCluster] Remove manually: '{}'".format(self._working_dir))
 			return True
 		else:
 			self._logger.error("[MGCluster] Clustering with odd result, several files found!")
-			self._logger.warning("[MGCluster] Remove manually: '{}".format(self._working_dir))
+			self._logger.warning("[MGCluster] Remove manually: '{}'".format(self._working_dir))
 			return False
 
-	def _get_symbolic_link_path(self, marker_gene_fasta):
-		basename = os.path.basename(marker_gene_fasta)
+	def _get_symbolic_link_path(self, original_file_path):
+		basename = os.path.basename(original_file_path)
 		new_path = os.path.join(self._working_dir, basename)
-		os.symlink(marker_gene_fasta, new_path)
+		os.symlink(original_file_path, new_path)
 		return new_path
 
-	def _get_mothur_cmd(self, marker_gene_fasta, cutoff):
+	def _get_mothur_cmd(self, marker_gene_fasta, cutoff, precision):
 		basename = os.path.basename(marker_gene_fasta)
 		filename, extension = os.path.splitext(basename)
 		mothur_cmd = """unique.seqs(fasta={mg_fasta})
 align.seqs(candidate=current, template={ref_align}, align=gotoh, flip=t, processors={processors})
 remove.seqs(accnos={filename}.unique.flip.accnos, fasta=current, name=current)
-dist.seqs(oldfasta={ref_align}, column={local_dist}, cutoff={cutoff}, processors={processors}, calc=onegap, countends=F)
 merge.files(input={filename}.names-{ref_names}, output={filename}.merged.names)
 merge.files(input={filename}.pick.names-{ref_names}, output={filename}.merged.names)
 set.current(name={filename}.merged.names, column={local_dist})
-cluster(cutoff={cutoff}, method=furthest, precision=100)"""
-		return mothur_cmd.format(filename=filename,
-								 mg_fasta=marker_gene_fasta,
+dist.seqs(oldfasta={ref_align}, column=current, cutoff={cutoff}, processors={processors}, calc=onegap, countends=F)
+set.current(column={local_dist})
+cluster(cutoff={cutoff}, method=furthest, precision={precision})"""
+		return mothur_cmd.format(wd=self._working_dir,
+								 debug=self._working_dir,
+								 #filename=os.path.join(self._working_dir, filename),
+								 filename=filename,
+								 mg_fasta=basename,
 								 ref_align=self._ref_silva_alignment,
 								 ref_names=self._ref_silva_names,
 								 local_dist=self._local_distance,
 								 processors=self._max_processors,
-								 cutoff=cutoff)
+								 cutoff=cutoff,
+								 precision=precision)
